Download files using cached revision id
This should ensure consistent downloads, since it was previously
possible that the cached MD5 sum of a recently changed object was used
as the chunk cache / request id for the download. By also using the
cached revision id, this should no longer be a problem.

Since for some reason the v3 revisions API does not allow HTTP range
requests, we switch to the v2 API download URLS, which work fine.